Move istiod patch into hook
